Do not allow other users to edit or delete my posts (#12)
do not allow other users to edit or delete my posts

+++ lib/finsta_web/live/post_live/index.ex
@@ -29,9 +29,17 @@ defmodule FinstaWeb.PostLive.Index do
   end
 
   defp apply_action(socket, :edit, %{"id" => id}) do
-    socket
-    |> assign(:page_title, "Edit Post")
-    |> assign(:post, Posts.get_post!(id))
+    post = Posts.get_post!(id)
+
+    if post.user_id == socket.assigns.current_user.id do
+      socket
+      |> assign(:page_title, "Edit Post")
+      |> assign(:post, post)
+    else
+      socket
+      |> put_flash(:error, "You are not authorized to edit this post")
+      |> redirect(to: ~p"/posts")
+    end
   end
 
   defp apply_action(socket, :new, _params) do
@@ -66,8 +74,15 @@ defmodule FinstaWeb.PostLive.Index do
   @impl true
   def handle_event("delete", %{"id" => id}, socket) do
     post = Posts.get_post!(id)
-    {:ok, _} = Posts.delete_post(post)
 
-    {:noreply, stream_delete(socket, :posts, post)}
+    if post.user_id == socket.assigns.current_user.id do
+      {:ok, _} = Posts.delete_post(post)
+
+      {:noreply, stream_delete(socket, :posts, post)}
+    else
+      socket
+      |> put_flash(:error, "You are not authorized to delete this post")
+      |> redirect(to: ~p"/posts")
+    end
   end
 end
